[REF] add upload image for story & fix update resource file detect file type auto
