+++ apps/storage/src/constant/storage-patterns.constant.ts
@@ -5,3 +5,8 @@ export const STORAGE_MESSAGE_PATTERNS = {
     CREATE: 'storage.create',
     UPLOAD: 'storage.upload'
 };
+
+
+export const STORAGE_MESSAGE_DB_RESPONSE = {
+    WRONG_UPLOAD: 'storage.can-not-replace-wrong-type-of-file',
+  }
